+++ Packages/ai.natml.vision.blazepalm/Runtime/BlazePalmPipeline.cs
@@ -1,12 +1,13 @@
 /* 
 *   BlazePalm
-*   Copyright (c) 2022 NatML Inc. All Rights Reserved.
+*   Copyright © 2023 NatML Inc. All Rights Reserved.
 */
 
 namespace NatML.Vision {
 
     using System;
     using System.Collections.Generic;
+    using System.Threading.Tasks;
     using UnityEngine;
     using NatML.Features;
     using NatML.Internal;
@@ -18,52 +19,29 @@ namespace NatML.Vision {
 
         #region --Client API--
         /// <summary>
-        /// Create the BlazePalm pipeline.
-        /// </summary>
-        /// <param name="detector">BlazePalm detector model data.</param>
-        /// <param name="predictor">BlazePalm landmark predictor model data.</param>
-        /// <param name="maxDetections">Maximum number of detections to return.</param>
-        public BlazePalmPipeline (MLModelData detector, MLModelData predictor, int maxDetections = Int32.MaxValue) {
-            this.detectorData = detector;
-            this.predictorData = predictor;
-            this.detectorModel = new MLEdgeModel(detector);
-            this.predictorModel = new MLEdgeModel(predictor);
-            this.detector = new BlazePalmDetector(detectorModel);
-            this.predictor = new BlazePalmPredictor(predictorModel);
-            this.maxDetections = maxDetections;
-        }
-
-        /// <summary>
         /// Detect hands in an image.
         /// </summary>
         /// <param name="inputs">Input image. This MUST be an `MLImageFeature`.</param>
         /// <returns>Detected hands.</returns>
         public BlazePalmPredictor.Hand[] Predict (params MLFeature[] inputs) {
-            // Check
-            if (inputs.Length != 1)
-                throw new ArgumentException(@"BlazePalm pipeline expects a single feature", nameof(inputs));
             // Check type
-            var input = inputs[0] as MLImageFeature;
-            if (input == null)
+            if (!(inputs[0] is MLImageFeature imageFeature))
                 throw new ArgumentException(@"BlazePalm pipeline expects an image feature", nameof(inputs));
-            // Detect poses
-            (input.mean, input.std) = detectorData.normalization;
-            input.aspectMode = detectorData.aspectMode;
-            var detections = detector.Predict(input);
+            // Detect ROIs
+            var detections = detector.Predict(imageFeature);
             // Predict landmarks
             var capacity = Mathf.Min(detections.Length, maxDetections);
             var result = new List<BlazePalmPredictor.Hand>(capacity);
             for (var i = 0; i < capacity; ++i) {
                 // Extract ROI
                 var detection = detections[i];
-                var roi = input.RegionOfInterest(detection.regionOfInterest, -detection.rotation, Color.black);
+                var roi = detection.regionOfInterest;
+                var roiFeature = new MLImageFeature(roiBuffer, roi.width, roi.height);
+                imageFeature.CopyTo(roiFeature, roi, -detection.rotation, Color.black);
                 // Predict landmarks
-                (roi.mean, roi.std) = predictorData.normalization;
-                roi.aspectMode = predictorData.aspectMode;
-                var landmarks = predictor.Predict(roi);
+                var landmarks = predictor.Predict(roiFeature);
                 // Create pose
-                var inputType = predictorModel.inputs[0] as MLImageType;
-                var keypoints = new BlazePalmPredictor.Keypoints(landmarks.keypoints.data, inputType, detection.regionOfInterestToImageMatrix);
+                var keypoints = new BlazePalmPredictor.Keypoints(landmarks.keypoints.data, landmarks.keypoints.imageType, detection.regionOfInterestToImageMatrix);
                 var hand = new BlazePalmPredictor.Hand(landmarks.score, landmarks.handedness, keypoints, landmarks.keypoints3D);
                 result.Add(hand);
             }
@@ -75,20 +53,48 @@ namespace NatML.Vision {
         /// Dispose the pipeline and release resources.
         /// </summary>
         public void Dispose () {
-            detectorModel.Dispose();
-            predictorModel.Dispose();
+            detector.Dispose();
+            predictor.Dispose();
+        }
+
+        /// <summary>
+        /// Create the BlazePalm pipeline.
+        /// </summary>
+        /// <param name="minScore">Minimum hand candidate score.</param>
+        /// <param name="maxDetections">Maximum number of detections to return.</param>
+        /// <param name="configuration">Edge model configuration.</param>
+        /// <param name="accessKey">NatML access key.</param>
+        public static async Task<BlazePalmPipeline> Create (
+            float minScore = 0.5f,
+            int maxDetections = Int32.MaxValue,
+            MLEdgeModel.Configuration configuration = null,
+            string accessKey = null
+        ) {
+            // Create detector and predictor
+            const float MaxIoU = 0.5f;
+            var detectorTask = BlazePalmDetector.Create(minScore, MaxIoU, configuration, accessKey);
+            var predictorTask = BlazePalmPredictor.Create(configuration, accessKey);
+            // Wait until created
+            await Task.WhenAll(detectorTask, predictorTask);
+            // Create pipeline
+            var pipeline = new BlazePalmPipeline(detectorTask.Result, predictorTask.Result, maxDetections);
+            return pipeline;
         }
         #endregion
 
 
         #region --Operations--
-        private readonly MLModelData detectorData;
-        private readonly MLModelData predictorData;
-        private readonly MLEdgeModel detectorModel;
-        private readonly MLEdgeModel predictorModel;
         private readonly BlazePalmDetector detector;
         private readonly BlazePalmPredictor predictor;
         private readonly int maxDetections;
+        private readonly byte[] roiBuffer; // this is a memory optimization to prevent allocations when extracting ROIs
+
+        private BlazePalmPipeline (BlazePalmDetector detector, BlazePalmPredictor predictor, int maxDetections = Int32.MaxValue) {
+            this.detector = detector;
+            this.predictor = predictor;
+            this.maxDetections = maxDetections;
+            this.roiBuffer = new byte[640 * 480 * 4]; // just has to be big enough for the largest possible ROI
+        }
         #endregion
     }
 }